Move rl_trainer into separate file
